Rename optee_ta_security to ta_security_api
