+++ trusted_application_demo/client_application/aes_certify_messaging/publisher/CMakeLists.txt
@@ -20,13 +20,14 @@ include_directories(${OPTEE_CLIENT_EXPORT}/include)
 
 find_package(ament_cmake REQUIRED)
 find_package(rclcpp REQUIRED)
+find_package(rcutils REQUIRED)
 find_package(aes_custom_interface REQUIRED)
 find_package(ca_security_api REQUIRED)
 include_directories(${ca_security_api_INCLUDE_DIRS}/ca_security_api)
 
-find_package(optee_ta_security REQUIRED)
-ament_export_dependencies(optee_ta_security)
-include_directories(${optee_ta_security_INCLUDE_DIRS}/optee_ta_security)
+find_package(ta_security_api REQUIRED)
+ament_export_dependencies(ta_security_api)
+include_directories(${ta_security_api_INCLUDE_DIRS}/ta_security_api)
 
 file(GLOB_RECURSE SRC_FILES_CXX ${PROJECT_SOURCE_DIR}/*.cpp)
