+++ trusted_application_demo/client_application/aes_certify_messaging/publisher/package.xml
@@ -10,9 +10,10 @@
   <buildtool_depend>ament_cmake</buildtool_depend>
 
   <depend>rclcpp</depend>
+  <depend>rcutils</depend>
   <depend>aes_custom_interface</depend>
   <depend>ca_security_api</depend>
-  <depend>optee_ta_security</depend>
+  <depend>ta_security_api</depend>
 
   <export>
     <build_type>ament_cmake</build_type>
