+++ trusted_application_demo/client_application/aes_certify_messaging/subscriber/package.xml
@@ -12,7 +12,7 @@
   <depend>rclcpp</depend>
   <depend>aes_custom_interface</depend>
   <depend>ca_security_api</depend>
-  <depend>optee_ta_security</depend>
+  <depend>ta_security_api</depend>
 
   <export>
     <build_type>ament_cmake</build_type>
