+++ trusted_application_demo/ta_security_api/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="2">
-  <name>optee_ta_security</name>
+  <name>ta_security_api</name>
   <version>0.0.1</version>
   <description>Trusted Application running in TEE</description>
   <maintainer email="florian.depraz@arm.com">Florian Depraz</maintainer>
@@ -13,4 +13,3 @@
     <build_type>ament_cmake</build_type>
   </export>
 </package>
-
