+++ trusted_application_demo/ta_security_api/src/Makefile
